+++ lib/src/advanced_grouped_list_order.dart
@@ -1,5 +1,5 @@
-/// Used to define the order of a [StickyGroupedListView].
-/// Defines the order in which items are grouped in a [StickyGroupedListView].
+/// Used to define the order of a [AdvancedGroupedListView].
+/// Defines the order in which items are grouped in a [AdvancedGroupedListView].
 enum AdvancedGroupedListOrder {
   /// Ascending order.
   asc,