+++ lib/src/advanced_grouped_list_widget.dart
@@ -120,7 +120,7 @@ class AdvancedGroupedListView<T, E> extends StatefulWidget {
   final bool addSemanticIndexes;
 
   /// The minimum cache extent used by the underlying scroll lists.
-  /// See [ScrollView.cacheExtent].
+  /// See [ScrollView.scrollCacheExtent].
   ///
   /// Note that the [ScrollablePositionedList] uses two lists to simulate long
   /// scrolls, so using the [ScrollController.scrollTo] method may result